fix: remove unused Compiler and Runner classes; streamline compilation process in CphNg
