+++ src/utils/exec.ts
@@ -1,0 +1,30 @@
+// Copyright (C) 2025 Langning Chen
+//
+// This file is part of cph-ng.
+//
+// cph-ng is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// cph-ng is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with cph-ng.  If not, see <https://www.gnu.org/licenses/>.
+
+import { promisify } from 'util';
+import { exec } from 'child_process';
+import { access } from 'fs/promises';
+
+export const execAsync = promisify(exec);
+export const exists = async (path: string): Promise<boolean> => {
+    try {
+        await access(path);
+        return true;
+    } catch {
+        return false;
+    }
+};
